# Core: Listen for test provider events before rendering test providers